Make implicit casts explicit (to silence xcode warnings)

+++ dataio/FileDescriptorDataIO.cpp
@@ -43,7 +43,7 @@ int32 FileDescriptorDataIO :: Read(void * buffer, uint32 size)
    int fd = _fd.GetFileDescriptor();
    if (fd >= 0)
    {
-      int r = read_ignore_eintr(fd, buffer, size);
+      int r = (int)read_ignore_eintr(fd, buffer, size);
       return _blocking ? r : ConvertReturnValueToMuscleSemantics(r, size, _blocking);
    }
    else return -1;
@@ -54,7 +54,7 @@ int32 FileDescriptorDataIO :: Write(const void * buffer, uint32 size)
    int fd = _fd.GetFileDescriptor();
    if (fd >= 0)
    {
-      int w = write_ignore_eintr(fd, buffer, size);
+      int w = (int)write_ignore_eintr(fd, buffer, size);
       return _blocking ? w : ConvertReturnValueToMuscleSemantics(w, size, _blocking);
    }
    else return -1;
